yash: test: compare_integers: intmax_t で比較するつもりが int で比較していた yash: variable: get_variable: 位置パラメータインデックスの型を uintmax_t に変更
git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@1156 048f04df-13f5-43d7-8114-9f9ceecaec24

+++ builtins/test.c
@@ -380,7 +380,7 @@ bool is_term_delimiter(const char *word)
  * the second respectively. */
 int compare_integers(const char *word1, const char *word2)
 {
-    int i1, i2;
+    intmax_t i1, i2;
     char *end;
 
     errno = 0;
